fixing modal flash issue, create note data now getting captured
