+++ app/javascript/lib/types/NoteRQ.ts
@@ -1,4 +1,7 @@
+import {Tag} from "./Tag";
+
 export interface NoteRQ {
     title: string;
     content: string;
+    tags: Tag[]
 }
