+++ app/javascript/lib/types/Tag.ts
@@ -1,6 +1,6 @@
 export interface Tag {
-    id: number;
+    id?: number;
     name: string;
-    created_at: string;
-    updated_at: string;
+    created_at?: string;
+    updated_at?: string;
 }
